fixes crash for large PSBs
- here is a test image: https://www.spacetelescope.org/static/archives/images/original/heic1502a.psb
- by casting, we don't get an overflow when computing the bytesPerChannel

+++ ImageLounge/3rdparty/libqpsd/qpsdhandler.cpp
@@ -105,8 +105,8 @@ bool QPsdHandler::read(QImage *image)
 
     input >> compression;
 
-    quint64 totalBytesPerChannel = width * height * depth / 8;
-    quint64 size = channels * totalBytesPerChannel;
+    quint64 totalBytesPerChannel = (quint64)width * height * depth / 8;
+    quint64 size = (quint64)channels * totalBytesPerChannel;
     QByteArray imageData;
 
     switch (compression) {
